+++ src/context/values.ts
@@ -75,11 +75,6 @@ export enum ContextValues {
   /** The user applied a search string to the Flink Artifacts view. */
   flinkArtifactsSearchApplied = "confluent.flinkArtifactsSearchApplied",
   /**
-   * PREVIEW: Are Flink resources and associated actions visible?
-   * (This should go away once the `confluent.preview.enableFlink` setting is removed.)
-   */
-  flinkEnabled = "confluent.flinkEnabled",
-  /**
    * EXPERIMENTAL: Is the chat participant enabled?
    * (This should go away once the `confluent.experimental.enableChatParticipant` setting is removed.)
    */